rebase onto 0.10.29
