+++ wandb/sdk/internal/file_stream.py
@@ -5,6 +5,7 @@ import itertools
 import logging
 import os
 import sys
+import random
 import requests
 import threading
 import time
@@ -236,6 +237,7 @@ class FileStreamApi(object):
         posted_anything_time = time.time()
         ready_chunks = []
         finished = None
+        last_log = 0
         while finished is None:
             items = self._read_queue()
             for item in items:
@@ -258,14 +260,14 @@ class FileStreamApi(object):
             if cur_time - posted_anything_time > self.heartbeat_seconds:
                 posted_anything_time = cur_time
                 self._handle_response(
-                    util.request_with_retry(
+                    request_with_retry(
                         self._client.post,
                         self._endpoint,
                         json={"complete": False, "failed": False},
                     )
                 )
         # post the final close message. (item is self.Finish instance now)
-        util.request_with_retry(
+        request_with_retry(
             self._client.post,
             self._endpoint,
             json={"complete": True, "exitcode": int(finished.exitcode)},
@@ -274,9 +276,9 @@ class FileStreamApi(object):
     def _handle_response(self, response):
         """Logs dropped chunks and updates dynamic settings"""
         if isinstance(response, Exception):
-            raise response
             wandb.termerror("Droppped streaming file chunk (see wandb/debug.log)")
             logging.error("dropped chunk %s" % response)
+            raise response
         elif response.json().get("limits"):
             parsed = response.json()
             self._api.dynamic_settings.update(parsed["limits"])
@@ -297,7 +299,7 @@ class FileStreamApi(object):
 
         for fs in file_stream_utils.split_files(files, max_mb=10):
             self._handle_response(
-                util.request_with_retry(
+                request_with_retry(
                     self._client.post,
                     self._endpoint,
                     json={"files": fs},
@@ -330,3 +332,87 @@ class FileStreamApi(object):
         """
         self._queue.put(self.Finish(exitcode))
         self._thread.join()
+
+
+MAX_SLEEP_SECONDS = 60 * 5
+
+
+def request_with_retry(func, *args, **kwargs):
+    """Perform a requests http call, retrying with exponential backoff.
+
+    Arguments:
+        func: An http-requesting function to call, like requests.post
+        max_retries: Maximum retries before giving up. By default we retry 30 times in ~2 hours before dropping the chunk
+        *args: passed through to func
+        **kwargs: passed through to func
+    """
+    max_retries = kwargs.pop("max_retries", 30)
+    retry_callback = kwargs.pop("retry_callback", None)
+    sleep = 2
+    retry_count = 0
+    while True:
+        try:
+            try:
+                response = func(*args, **kwargs)
+                response.raise_for_status()
+                return response
+            except (
+                requests.exceptions.ConnectionError,
+                requests.exceptions.HTTPError,
+                requests.exceptions.Timeout,
+            ) as e:
+                if isinstance(e, requests.exceptions.HTTPError):
+                    # Non-retriable HTTP errors.
+                    #
+                    # We retry 500s just to be cautious, and because the back end
+                    # returns them when there are infrastructure issues. If retrying
+                    # some request winds up being problematic, we'll change the
+                    # back end to indicate that it shouldn't be retried.
+                    if (
+                        e.response is not None
+                        and e.response.status_code in {400, 403, 404, 409}
+                    ) or (
+                        e.response is not None
+                        and e.response.status_code == 500
+                        and e.response.content
+                        == b'{"error":"context deadline exceeded"}\n'
+                    ):
+                        return e
+
+                if retry_count == max_retries:
+                    return e
+                retry_count += 1
+                delay = sleep + random.random() * 0.25 * sleep
+                if isinstance(e, requests.exceptions.HTTPError) and (
+                    e.response is not None and e.response.status_code == 429
+                ):
+                    err_str = "Filestream rate limit exceeded, retrying in {} seconds".format(
+                        delay
+                    )
+                    if retry_callback:
+                        retry_callback(e.response.status_code, err_str)
+                    logger.info(err_str)
+                else:
+                    pass
+                    logger.warning(
+                        "requests_with_retry encountered retryable exception: %s. func: %s, args: %s, kwargs: %s",
+                        e,
+                        func,
+                        args,
+                        kwargs,
+                    )
+                time.sleep(delay)
+                sleep *= 2
+                if sleep > MAX_SLEEP_SECONDS:
+                    sleep = MAX_SLEEP_SECONDS
+            except requests.exceptions.RequestException as e:
+                logger.error(response.json()["error"])  # XXX clean this up
+                logger.exception(
+                    "requests_with_retry encountered unretryable exception: %s", e
+                )
+                return e
+        except Exception as e:
+            logger.exception(
+                "Filestream encountered unhandled exception in retry: {}".format(e)
+            )
+            raise e
